tests_app/requirements.txt
nose
django-nose
mock
ipdb
pymongo